\name{hello}
\alias{hello}
\title{Hello, World!}
\usage{
getchart()
}
\description{
Call stored data stored in google spreadsheets and raw bar plot to get known current supplying number all over the country.
This is a simple demonstration to develope more detailed model in the future.
}
\examples{
getchart()
}
